build errors fixed - ready to deploy
