+++ apps/web/src/components/ErrorContent/style.css
