+++ apps/web/src/components/UserTable/page.tsx
@@ -23,7 +23,6 @@ export default function UserTable({ role }: UserTableProps) {
         const data = await userRoleService.getUsersByRole(role);
         setUsers(data);
       } catch (err) {
-        console.error('Using fallback mock data', err);
         setError(axios.isAxiosError(err) ? err.message : 'An unknown error occurred');
       } finally {
         setLoading(false);
